src/requirements.txt
fastapi[all]==0.104.1
jinja2==3.1.2
uvicorn[standard]==0.24.0.post1
python-dotenv==1.0.0
celery[redis]==5.3.6
aiomysql==0.2.0
pymysql==1.1.0
requests==2.31.0
